# backend/docs/openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "Books API",
    "version": "1.0",
    "description": "This is a sample server for managing books.",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "API Support",
      "url": "http://www.swagger.io/support",
      "email": "support@swagger.io"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
    "/books": {
      "get": {
        "summary": "Get all books",
        "tags": [
          "books"
        ],
        "description": "Get details of all books",
        "operationId": "getAllBooks",
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Books"
              }
            }
          },
          "500": {
            "description": "Failed to fetch books",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/{id}": {
      "get": {
        "summary": "Get book by ID",
        "tags": [
          "books"
        ],
        "description": "Get details of a book by ID",
        "operationId": "getBookByID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "Book ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "$ref": "#/definitions/Books"
            }
          },
          "404": {
            "description": "Book not found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "put": {
        "summary": "Update a book",
        "tags": [
          "books"
        ],
        "description": "Update details of an existing book by ID",
        "operationId": "updateBook",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "Book ID"
          },
          {
            "in": "body",
            "name": "book",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Books"
            },
            "description": "Book details to update"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "$ref": "#/definitions/Books"
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Book not found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Delete a book",
        "tags": [
          "books"
        ],
        "description": "Delete a book by ID",
        "operationId": "deleteBook",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64",
            "description": "Book ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Book deleted successfully"
                }
              }
            }
          },
          "404": {
            "description": "Book not found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/process-url": {
      "post": {
        "summary": "Process URL",
        "tags": [
          "url"
        ],
        "description": "Process URL based on specified operation",
        "operationId": "handleURLProcessing",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/URLRequest"
            },
            "description": "URL request"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "$ref": "#/definitions/URLResponse"
            }
          },
          "400": {
            "description": "Invalid JSON input",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    }
  },
  "definitions": {
    "Books": {
      "type": "object",
      "properties": {
        "SSID": {
          "type": "integer",
          "format": "int64"
        },
        "Title": {
          "type": "string",
          "maxLength": 255
        },
        "Author": {
          "type": "string",
          "maxLength": 255
        },
        "Year": {
          "type": "string",
          "maxLength": 255
        }
      }
    },
    "URLRequest": {
      "type": "object",
      "properties": {
        "URL": {
          "type": "string"
        },
        "Operation": {
          "type": "string",
          "enum": [
            "canonical",
            "redirection",
            "all"
          ]
        }
      }
    },
    "URLResponse": {
      "type": "object",
      "properties": {
        "ProcessedURL": {
          "type": "string"
        }
      }
    }
  }
}